gui/requirements.txt
Flask==2.3.3
PyYAML==6.0.1
Jinja2==3.1.4
